pyproject.toml
[tool.poetry]
name = "intimacoes"
version = "0.1.0"
description = ""
authors = ["Alexandre Ito <shiguemialexandre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
bcrypt = "^4.2.0"
extra-streamlit-components = "^0.1.71"
openai = "^1.37.1"
nltk = "^3.8.1"
scikit-learn = "^1.5.1"
python-magic-bin = "^0.4.14"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
